Refactor getWebpackBuildPath.js into multiple files
