+++ src/server/Html.js
@@ -5,9 +5,9 @@ import {RouterContext} from 'react-router';
 import {renderToString} from 'react-dom/server';
 import makeSegmentSnippet from '@segment/snippet';
 import {auth0} from 'universal/utils/clientOptions';
-import getWebpackBuildPath from 'universal/utils/getWebpackBuildPath';
+import getWebpackPublicPath from 'server/utils/getWebpackPublicPath';
 
-const webpackPublicPath = getWebpackBuildPath();
+const webpackPublicPath = getWebpackPublicPath();
 const segKey = process.env.SEGMENT_WRITE_KEY;
 const segmentSnippet = segKey && makeSegmentSnippet.min({
   host: 'cdn.segment.com',
